requirements.txt
absl-py==0.7.1
alabaster==0.7.10
altair==2.0.1
annoy==1.11.5
asn1crypto==0.22.0
astor==0.8.0
astroid==1.5.3
astropy==2.0.2
astunparse==1.6.3
awscli==1.15.7
azure-common==1.1.25
azure-nspkg==3.0.2
azure-storage==0.36.0
Babel==2.5.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcolz==1.2.1
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
bleach==1.5.0
blis==0.4.1
bokeh==0.12.10
boto==2.48.0
boto3==1.7.7
botocore==1.10.7
Bottleneck==1.2.1
bz2file==0.98
cachetools==2.0.1
catalogue==1.0.0
certifi==2018.4.16
cffi==1.10.0
chardet==3.0.4
cld2-cffi==0.1.4
click==6.7
cliff==2.8.1
cloudpickle==0.4.0
cmd2==0.8.5
colorama==0.3.7
comment-parser==1.1.6
configparser==3.5.0
contextlib2==0.5.5
cryptography==2.3
cssselect==1.0.3
cycler==0.10.0
cymem==2.0.3
Cython==0.26.1
cytoolz==0.9.0.1
dask==0.17.2
decorator==4.3.0
dill==0.2.7.1
distributed==1.19.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.14
dpu-utils==0.2.17
elasticsearch==7.0.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.0/en_core_web_sm-2.3.0.tar.gz
entrypoints==0.2.3
enum34==1.1.10
et-xmlfile==1.0.1
fastai==0.6
fastcache==1.0.2
feather-format==0.4.0
filelock==2.0.12
Flask==1.0
Flask-Cors==3.0.3
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-2.3.0/fr_core_news_sm-2.3.0.tar.gz
ftfy==4.4.3
future==0.16.0
fuzzywuzzy==0.18.0
gast==0.2.2
gensim==3.4.0
gevent==1.2.2
gitdb==4.0.5
GitPython==3.1.3
glob2==0.5
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
gql==0.2.0
graphql-core==1.1
graphviz==0.8.2
greenlet==0.4.12
grpcio==1.22.0
h5py==2.9.0
HeapDict==1.0.0
hnswlib==0.4.0
html5lib==0.9999999
idna==2.6
ijson==3.1.post0
importlib-metadata==1.6.1
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isoweek==1.3.3
itsdangerous==0.24
javalang==0.13.0
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
joblib==0.15.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
kaggle-cli==0.12.13
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
ktext==0.40
lazy-object-proxy==1.3.1
lxml==4.2.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.2
maven==0.0.6
MechanicalSoup==0.8.0
mistune==0.8.3
mizani==0.4.6
more-itertools==8.4.0
msgpack==1.0.0
msgpack-numpy==0.4.6.post0
msgpack-python==0.5.6
multiprocess==0.70.10
murmurhash==1.0.2
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.4
nltk==3.5
nmslib==2.0.6
notebook==5.7.8
numexpr==2.7.1
numpy==1.16.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.45.1
opencv-python==3.4.0.12
opt-einsum==3.2.1
palettable==3.1.0
pandas==0.23.0
pandas-summary==0.0.7
pandocfilters==1.4.2
parso==0.2.1
pathos==0.2.6
pathtools==0.1.2
patsy==0.5.0
pbr==4.0.2
pexpect==4.5.0
pickleshare==0.7.4
Pillow==6.2.0
plac==1.1.3
plotnine==0.3.0
pox==0.2.8
ppft==1.6.6.2
preshed==3.0.2
prettytable==0.7.2
progressbar2==3.34.3
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==1.0.15
protobuf==3.9.1
psutil==5.6.6
ptyprocess==0.5.2
py==1.4.34
pyarrow==0.9.0
pyasn1==0.4.2
pyasn1-modules==0.2.8
pybind11==2.5.0
pycparser==2.18
pyemd==0.5.1
Pygments==2.2.0
pygraphviz==1.5
pyparsing==2.2.0
pyperclip==1.6.0
Pyphen==0.9.5
pytest==3.2.1
python-dateutil==2.7.3
python-Levenshtein==0.12.0
python-magic==0.4.15
python-utils==2.3.0
pytz==2018.4
PyYAML==5.1
pyzmq==17.0.0
qtconsole==4.3.1
regex==2020.6.8
requests==2.20.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.1
seaborn==0.8.1
Send2Trash==1.5.0
sentencepiece==0.1.91
sentry-sdk==0.15.1
SetSimilaritySearch==0.1.7
shortuuid==1.0.1
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
sklearn-pandas==1.6.0
smart-open==1.5.7
smmap==3.0.4
sortedcontainers==1.5.7
spacy==2.3.0
srsly==1.0.2
statsmodels==0.8.0
stevedore==1.28.0
subprocess32==3.5.4
table==0.0.4
tables==3.6.1
tblib==1.3.2
tensorboard==1.12.2
tensorboard-plugin-wit==1.7.0
tensorboardX==2.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.4.0
tensorflow-hub==0.5.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
textacy==0.6.2
thinc==7.4.1
threadpoolctl==2.1.0
toolz==0.9.0
torch==0.4.1
torchtext==0.6.0
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.2
tree-sitter==0.1.1
typing==3.6.4
Unidecode==1.1.1
unrar==0.4
urllib3==1.24.3
wandb==0.9.1
wasabi==0.7.0
watchdog==0.10.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
wget==3.2
widgetsnbextension==3.2.1
wrapt==1.11.2
xlrd==1.2.0
xmltodict==0.12.0
zict==0.1.3
zipp==3.1.0
